+++ src/Providers/PbMiddlewareServiceProvider.php
@@ -4,6 +4,7 @@ namespace Anibalealvarezs\Projectbuilder\Providers;
 
 use Anibalealvarezs\Projectbuilder\Helpers\AeasHelpers as AeasHelpers;
 use Illuminate\Contracts\Container\BindingResolutionException;
+use Illuminate\Contracts\Http\Kernel;
 use Illuminate\Support\ServiceProvider;
 
 class PbMiddlewareServiceProvider extends ServiceProvider
@@ -13,22 +14,21 @@ class PbMiddlewareServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function boot(\Illuminate\Contracts\Http\Kernel $kernel)
+    public function boot(Kernel $kernel)
     {
-        $aeas = new AeasHelpers();
         //global middleware
-        $kernel->prependMiddleware($aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'HttpsMiddleware');
-        $kernel->pushMiddleware($aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'HttpsMiddleware');
-        $kernel->prependMiddleware($aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'SingleSessionMiddleware');
-        $kernel->pushMiddleware($aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'SingleSessionMiddleware');
+        $kernel->prependMiddleware(AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'HttpsMiddleware');
+        $kernel->pushMiddleware(AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'HttpsMiddleware');
+        $kernel->prependMiddleware(AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'SingleSessionMiddleware');
+        $kernel->pushMiddleware(AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'SingleSessionMiddleware');
         //router middleware
         $router = $this->app['router'];
-        $router->pushMiddlewareToGroup('web', $aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'HttpsMiddleware');
-        $router->aliasMiddleware('role_or_permission', $aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'RoleOrPermissionMiddleware');
-        $router->aliasMiddleware('is_user_viewable', $aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'IsUserViewableMiddleware');
-        $router->aliasMiddleware('is_user_editable', $aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'IsUserEditableMiddleware');
-        $router->aliasMiddleware('is_user_selectable', $aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'IsUserSelectableMiddleware');
-        $router->aliasMiddleware('is_user_deletable', $aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'IsUserDeletableMiddleware');
+        $router->pushMiddlewareToGroup('web', AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'HttpsMiddleware');
+        $router->aliasMiddleware('role_or_permission', AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'RoleOrPermissionMiddleware');
+        $router->aliasMiddleware('is_user_viewable', AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'IsUserViewableMiddleware');
+        $router->aliasMiddleware('is_user_editable', AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'IsUserEditableMiddleware');
+        $router->aliasMiddleware('is_user_selectable', AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'IsUserSelectableMiddleware');
+        $router->aliasMiddleware('is_user_deletable', AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'IsUserDeletableMiddleware');
     }
 
     /**
@@ -39,7 +39,6 @@ class PbMiddlewareServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $aeas = new AeasHelpers();
-        $this->app->make($aeas->vendor.'\\'.$aeas->package.'\Middleware\\'.$aeas->prefix.'HttpsMiddleware');
+        $this->app->make(AeasHelpers::AEAS_VENDOR.'\\'.AeasHelpers::AEAS_PACKAGE.'\Middleware\\'.AeasHelpers::AEAS_PREFIX.'HttpsMiddleware');
     }
 }